pyproject.toml
[tool.poetry]
name = "reviews-pipeline"
version = "0.1.0"
description = ""
authors = ["Nicolae <for.nicolae.popovici@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.36"
uuid = "^1.30"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120